refactor : 구조 변경
